Move auth handlers into auth module
These structs and the tests belong to the auth module because that is
where all the functionality is implemented.
